age_gender_detection/requirements.txt
absl-py==0.7.1
appdirs==1.4.3
astor==0.7.1
bleach==1.5.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
cmake==3.18.4.post1
cos-python-sdk-v5==1.6.6
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
dicttoxml==1.7.4
dlib==19.18.0
easydict==1.9
enum34==1.1.6
feather-format==0.4.0
ffmpeg==1.4
Flask==1.0.2
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
graphviz==0.8.4
grpcio==1.25.0
h5py==2.9.0
hnswlib==0.3.4
html5lib==0.9999999
idna==2.8
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.10.1
joblib==0.13.2
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
llvmlite==0.28.0
Mako==1.1.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mmdnn==0.2.5
mock==2.0.0
msgpack==0.6.1
msgpack-numpy==0.4.4.2
mxnet==1.6.0
numba==0.43.0
numpy==1.17.3
oauthlib==3.1.0
onnx==1.5.0
opencv-contrib-python==4.2.0.32
opencv-python~=4.5.2.52
opt-einsum==3.0.1
pandas==0.24.2
pbr==5.1.3
Pillow==6.1.0
ply==3.11
protobuf==3.11.2
psutil==5.6.1
pyarrow==0.13.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pybind11==2.5.0
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.8.0
pytools==2019.1.1
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn~=0.24.2
scipy==1.4.1
six~=1.14.0
sklearn==0.0
slidingwindow==0.0.13
SQLAlchemy==1.3.3
tabulate==0.8.3
tensorboard==2.1.0
tensorflow-addons==0.8.3
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
tensorflow-model-optimization==0.1.3
tensorpack==0.9.1
termcolor==1.1.0
tf2onnx==1.5.2
tfp-nightly==0.9.0.dev20190926
thriftpy2==0.4.2
tqdm==4.31.1
typeguard==2.8.0
typing==3.7.4
typing-extensions==3.7.4
urllib3==1.24.1
Werkzeug==0.15.0
wrapt==1.11.1

mtcnn~=0.1.0